+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add `impl_from!` and `impl_into!` macros.
+
 ## 0.0.1 - 2022-09-25
 
 - Initial release with `impl_as_ref!`, `impl_as_mut!`, `impl_deref!`, and `impl_deref_mut!` macros.